refactor(admin): extract pure helpers to admin/util.js ES module (#1279 step 2/6)
Schritt 2/6 of the admin.js decomposition. Markus freigegeben: type="module".

- Extract the 6 step-1-tested pure helpers + REQUEST_STATUS_LABELS into a new
  ES module www/js/admin/util.js (escapeHtml, groupPlayersByPlatform,
  buildRequestRowHtml, _getAdminToken, _setAdminToken, _adminHeaders).
- Migrate `admin` from MINIFY to BUNDLES in scripts/build.mjs (format esm);
  admin.html now loads admin.min.js via <script type="module">.
- currentGame (admin-private mutable state) injected into util.js via a closure
  resolver registered once — no need to touch every assignment site.
- Compat shim: re-expose the extracted helpers on window so classic scripts
  loaded after the now-deferred admin module keep resolving them.
- Tests: drop the eval-based admin-helpers-loader.js, switch
  admin-helpers.test.js to real `import` from ../admin/util.js (200 green).
- npm run build + build:check drift-free; admin.min.js rebuilt.

Consumer load-order analysis (in PR body): no consumer reads an admin global at
parse time — party-lights/tts-settings have their own escapeHtml, wizard reads
window.loadStatus/loadSavedSettings/BeatifyHome only on a button click,
playlist-requests reads window.BEATIFY_VERSION at render time. All event-driven
→ migration is timing-safe.
